Extract `MockTokenAuthorizer` as an "official" mock.
